Add Ord instance for AnomaOpqaue using anomaEncode
It's possible to call anomaEncode on a value of AnomaOpaque becuase its
an Anoma builtin and at runtime it will either be a Nock Atom or a Nock
Cell, both of which can be arguments to anomaEncode.

logicToRef and logicFromRef are now classified as missing anoma
builtins. In the transparent instantitation of the RM these will be the
identity function.