Remove proving system trait
The current RM spec supports only compliance, resource logic and delta
proofs. We can model the proof records for each type of proof using a
ADT. We do not need a fully parametrized proving system trait to support
this.

In the future the spec may require us to support arbitrary proving
systems (i.e user-defined proving systems that can be fully parametrized
by their proof, instance, witness etc.). If this is the case we'll need
to reinstate the proving system trait.